requirements.txt
bottle
bottle_session
bottle_sqlalchemy
bottle_react==0.4.0
bottle_beaker
psycopg2
bcrypt
awesome-slugify
paste
